Fix issue with incorrect prices and missing quantity
The prices that are synced are always excluding tax. Now it will be
based on the configuration for the display of the prices in the
frontend.

Also the quantity of products wasn't sent, because the item was said
to be a order item, while it was in fact an invoice item. This has
been fixed now.