Implement debug mode and logging option and send document path (#2)
* Add document path as parameter for the Google Analytics call

When doing the call to Google Analytics, the document path was not
sent. This made the URL in Google Analytics set to "(not set)". This
has been added now, so the actual pageview is also tracked.

* Add debug option for server side synchronisation

* Fix issue with incorrect method call

In the previous commit the wrong method was used to check for debug
mode. This has been fixed now

* Change text to class constant

The text for the configuration setting to debug has been changed to
a class constant.

* Fix issue with incorrect prices and missing quantity

The prices that are synced are always excluding tax. Now it will be
based on the configuration for the display of the prices in the
frontend.

Also the quantity of products wasn't sent, because the item was said
to be a order item, while it was in fact an invoice item. This has
been fixed now.

* Enable logging of requests

It was already possible to log the debug response, but it wasn't
possible to log the requested URLs. This has now been implemented,
so that it's also possible to check the URLs that are called on a
production server (not recommended).